refactor: split ticket repository, fix f-string SQL and route DB access
Resolves PR #255 review issues #1, #2, #4:
- Split backend/db/repository/tickets.py (1031 lines) into tickets_core.py
  (CRUD) + tickets_stats.py (_TicketStatsMixin); TicketRepository inherits
  the mixin so its public API is unchanged. Removed the legacy ceiling entry
  from tools/check_modularity.py — new files stay under the 700-line limit.
- Replaced f-string SQL in get_open_tickets/get_ticket_count with safe
  string concatenation.
- Moved _get_cooldown_floor DB query into TicketRepository.get_cooldown_reset()
  so the internal API no longer calls Database.get_connection() directly
  (backend-first compliance).
